touches: Track and pass previous touch event too
The pending() handler now gets

  1. the starting state of the touch
  2. the immediately previous state of the touch
  3. the current end state of the touch

This allows for between-state calculations such as dy since the previous
event instead of just total dy.  I plan to use this in adapting the
volume control to be relative and limitless instead of absolute, which
is necessary when controlling the stereo receiver since there's no way
to get the current volume of it.