Soundman coordinates the stereo receiver and its various inputs
A unified remote!

The worst bit is that I can't (easily, if at all) get state from the
stereo receiver, so can't have the remote load up matching reality, and
thus have to resort to dead reckoning, of a sorts.  There's a little
more that could be easily done here to intuit reality from the input
states, which would be better (but still a heuristic).

The volume slider is the next worst bit, as it's inaccessible when
listening to Plink.  Need to think about this interface a bit, but I
think there's no avoiding a major UI change to make that better.